refactor(main.ts): :recycle: 抽离一些功能为vue插件
