+++ src/plugins/i18n.ts
@@ -1,0 +1,7 @@
+// 国际化
+import i18n from "@/lang/index";
+import type { App } from "vue";
+
+export function setupI18n(app: App<Element>) {
+  app.use(i18n);
+}
