+++ src/plugins/index.ts
@@ -1,0 +1,3 @@
+export * from "./icons";
+export * from "./i18n";
+export * from "./permission";
